requirements.txt
redis>=2.4.9
django>=1.7.0
django-modeldict==1.4.1
mock==1.0.1
django-waffle>=0.11
